mcp/agentic-workflow-server/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "agentic-workflow-server"
version = "0.1.0"
description = "MCP server for agentic-workflow state management and configuration"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "Agentic Workflow Contributors" }
]
keywords = ["mcp", "claude", "workflow", "agents"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "mcp>=1.0.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
]

[project.scripts]
agentic-workflow-server = "agentic_workflow_server.server:main"

[tool.hatch.build.targets.wheel]
packages = ["agentic_workflow_server"]

[tool.hatch.build]
include = [
    "agentic_workflow_server/**/*.py",
]
